pyproject.toml
[project]
name = "stockiq"
version = "0.0.1"
description = "Free S&P 500 Technical Analysis & Screener"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.50,<2.0",
    "yfinance>=0.2.30",
    "plotly>=5.0",
    "pandas>=2.0",
    "numpy>=1.20",
    "anthropic>=0.40.0",
    "groq>=0.9.0",
    "google-genai>=0.8.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "google-cloud-storage>=2.0",
]

[project.urls]
Homepage    = "https://github.com/Mamun/stockIQ"
Repository  = "https://github.com/Mamun/stockIQ"
"Bug Tracker" = "https://github.com/Mamun/stockIQ/issues"

[build-system]
requires      = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
python_files   = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

# ── ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length    = 120
exclude        = [".git", "__pycache__", ".streamlit"]

[tool.ruff.lint]
select = ["E", "W", "F"]
ignore = ["E501", "W291", "W293", "E402"]

# ── coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src"]
omit = [
    "src/stockiq/views/*",   # Streamlit page modules require a running server
    "src/stockiq/seo.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
